[ARM] igep0020 : Add support for IGEP v2.x rev. B boards
Note that this patch adds initial support for our IGEP v2.x board and really needs to be tested. This
increases VMALLOC_END to 0x28000000 and sets default PAGE_OFFSET for ARM architecture to 0xB0000000
instead of 0xC0000000, this prevents overlaps into virtual memory map.

I'm not sure about this solution so it needs to be checked.