ASOC: TWL4030 Audio capture fix
A couple of folks have noticed an issue with audio capture -- the
capture result is always silence.

The patch below is a quick fix for those with this issue.  There are
substantial changes to the codec driver that will be trickling down
from ASoC, and they deal with this issue differently.

So consider this as a bandaid for those who don't want to wait for the
trickle down :-)

Steve