serial: Auto RS485 direction control
To use correctly the RS485 interface chip we need to configure the serial port driver in the
kernel to manage the RTS signal (Request To Send) to enable the TX section (DE pin) only when
data are transmitted on TXD line (Transmit Data). This is a software only solution, requiring
no special hardware, but timing control for delays is only accurate to the next nearest jiffy.

Discussions on RS485 auto direction control seem to be split over a number of threads, see:
    *  http://marc.info/?t=121805745200001&r=1&w=2
    *  http://marc.info/?t=121690034200007&r=1&w=2
    *  http://marc.info/?t=122821576400003&r=1&w=2
    *  http://marc.info/?t=122822457300004&r=1&w=2
    *  http://marc.info/?t=122954923400001&r=1&w=2
    *  http://marc.info/?t=122964426200001&r=1&w=2
    *  http://marc.info/?t=121566595900001&r=1&w=2
    *  http://marc.info/?t=121805198200004&r=1&w=2