OMAP3: exp-base0010: Fix TSC2046 touchscreen detection
There is a regulador that controls a bunch of devices related to the
LCD like the touch controller, the backlight driver, and the powerup
display. If the LCD is not enabled the touchscreeen can't be detected
beacause the enable/disable_lcd functions turns on/off the regulator.
This is not what we want.

This patch turns on the regulator and does not stop.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>