ASoC: TWL4030: Disable REG_VOICE_IF by default
On IGEP boards, we need to set the TWL4030 PCM voice interface output
pins to high impedance, otherwise this module forces the level of
mcbsp3_fsx and mcbsp3_clkx pins, then the modem connected to uart2 on
IGEP boards doesn't work. As a workaround this patch disables the
REG_VOICE_IF by default.

PCM/codec/TDM/Bluetooth interface modes cannot be changed on the fly,
so we need to set the value 0x04 to TWL4030_REG_VOICE_IF at this file
	 -       0x00,  REG_VOICE_IF  (0xF)
	 +       0x04,  REG_VOICE_IF  (0xF)

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>