OMAP3: igep0020: Fix mux when using board.ei485 kernel parameter
Is too early for use omap_mux_init_signal function in early_param. The
mux_init call just fails with Could not set signal. This patch sets a
flag when board.ei485=no is passed in kernel cmdline, then uses this flag
to set the mux acording to the parameter in board init function.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>