omap3isp: isp: Re-enable ccdc vd1 IRQ
ISP has registers that are shadowed. This type of registers can be read and
written at any time but the written values don't take effect until some event
occurs, such as the start of a frame.

Since these registers can't be configured on the end-of-frame interrupt handler
we need another interrupt (VD1) at the middle of a frame to configure the ISP.

Signed-off-by: Javier Martinez Canillas <martinez.javier@gmail.com>