omap3isp: ispvideo: Decouple next buffer obtaining from last buffer releasing
The original ISP driver released the last buffer (waking up the pending process)
before returning the next buffer to the caller. Since we moved most of the logic
to the middle of the processing of a frame, the next buffer obtaining and the
last buffer releasing is made at different moments. So we have to decouple these
two actions.

Signed-off-by: Javier Martinez Canillas <martinez.javier@gmail.com>