OMAP3: igep00x0: Fix mux handling for UART2
OMAP3530 CBB package can have UART2 muxed on 2 pins: uart2_* and
mcbsp3_*.

On IGEP0020 the bluetooth uses uart2_* pins, if IGEP0022 board is
connected and we wont use the modem we should configure MODE1 for
mcbsp3_* pins.

On IGEP0030 the bluetooth uses mcbsp3_* pins, if BASE0010 rev. B
is connected and we wont use the modem we should configure MODE0
for uart2_* pins.

On IGEP0032 if BASE0010 rev. B is connected we should configure
MODE0 for uart2_* pins.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>