board-igep0020: force twl4030 codec registers reset at boot
If registers are not reset to the driver default values the
codec will have an unknown state.

In the igep specific case the issue is that if the codec voice interface
is disabled (REG_VOICE_IF = 0x04 in sound/soc/codec/twl4030.c) it will
never be set to that value and the expansion board UART2 will be
unusable.

Tested on IGEPv2 with expansion board rev.B

Unrelated cleanup: omap_pmic_init replaced with omap3_pmic_init

Signed-off-by: Enrico Butera <ebutera@users.berlios.de>
Signed-off-by: Javier Martinez Canillas <martinez.javier@gmail.com>