linux-rt: Add real-time support provided by Xenomai
Xenomai needs special kernel support to deliver fast and
deterministic response time to external interrupts, and also to
provide real-time services highly integrated with the standard Linux
kernel.

This support is provided by the Adeos real-time enabler [1], in the
form of a kernel patch [2] you have to apply against a kernel.

[1] http://www.gna.org/projects/adeos/
[2] adeos-ipipe-2.6.35.9-arm-1.18-03.patch

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>