OMAP3: igep0030: Fix power for dss_data[0..5] pins
On DM3730 seems dss_data[0..5] pins are powered via VDDS_SDI
power domain, on IGEP boards this domain is supplied by the
VPLL2 regulator. If the regulator is not enabled by the bootloader
then it will be enabled by the kernel. This patch ensures that
the regulator constraints are applied.

Note that on the AM/DM37x Hardware Reference Manual the power domain
for these pins is VDDS not VDDS_SDI, but configuring dss_data[0..5]
pins as gpio without VDDS_SDI doesn't work as expected.

The issue was detected using an IGEP COM MODULE + IGEP NEW YORK
board.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>