omap: serial: Fix bug when using rs485 mode.
After finishing to transmit in rs485 mode the UART_IER, UART_OMAP_SCR and
UART_MCR_RTS values weren't updated correctly.

This had the effect that once a transmission was made from a source node in
a semi-duplex communication, the transmitter didn't allow others nodes to
transmit anymore.

If the tty was closed after it finished the transmission and re-opened to
transmit again the bug wasn't spot. Only happened if the tty was kept opened.

Also, this patch make a few cleanups to the omap-serial driver.

We only have to update the Ready-to-Send control line if the user sets the
SER_RS485_RTS_ON_SEND flag and testing inside the function is redundant.

If the user didn't set the SER_RS485_RTS_ON_SEND flag, we don't have to
update the RTS line, since is the hardware who manages this.