omap3isp: Improve CCDC ISP driver to capture video in BT.656 data format.
This patch allows to correctly capture effective lines for each video frame
in interlaced data mode.
Also, the isp video driver walks through the pipeline to configure the sink
node according to the configuration of the source node pad.
So, application that are not media control framework aware can use the video
node as a normal V4L2 device.