OMAP3: base0010: Remove duplicate OMAP mux init for MCP251x CAN controller
Base0010 revision B init function set the OMAP mux pins for MCP251x
NRESET, IRQ and chip-select 0.

This is not necessary since the mux pins for this controller are already
configured previously.

Also, this duplication has a side effect on the MCP251x driver that fails
trying to reset the hardware on its probe function:

mcp251x spi1.0: MCP251x didn't enter in conf mode after reset
[   58.999420] mcp251x spi1.0: Probe failed
Cannot find device "can0"

So, this patch removes the duplicate to clean-up the code and fix this problem.

Signed-off-by: Javier Martinez Canillas <martinez.javier@gmail.com>