OMAP3: Restructure camera interface and TVP5151 support on IGEP boards
IGEP expansion boards igep0022 and base0010 include an TVP5151 video decoder
and a camera sensor can be attached to the IGEP COM module (igep0030).

This patch adds the camera interface configuration pins to each board OMAP
mux configuration so it is not necessary to configure this latter.

Also, since the video encoder is a feature of each expansion board, it makes
sense to move the TVP5150 ISP platform data to each board specific file and
use a generic camera initialization that is common to all expansion board.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>