board-igep00x0.c: Set max/min coordinates for touchscreen
The max/min values for the x/y coordinate of the touchscreen in
plattform data didn't reflect the real touchable area. As a result
it wasn't possible to reach the edges of the touch.
This patch changes the coordinates so that the whole touchscreen
area of the device is usable.

Signed-off-by: Matthias Brugger <mbrugger@iseebcn.com>