mmc: omap_hsmmc: Fix for bus width which improves SD card performance.
As a result of the forward port of the omap-hsmmc driver we introduced a
performance issue.

The MMC controller misses to use the platform data capabilities, some boards
uses pdata->caps to set the bus width and the driver ignores this. This patch
fixes this and improves the SD card performance.

  $ dd if=/dev/mmcblk0 of=/dev/null
  7882752+0 records in
  7882752+0 records out
  4035969024 bytes (4.0 GB) copied, 249.62 s, 16.2 MB/s

  $ dd if=/dev/zero of=/dev/mmcblk0
  7882753+0 records in
  7882752+0 records out
  4035969024 bytes (4.0 GB) copied, 1320.57 s, 3.1 MB/s

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>