usb: musb: ti81xx: Bypass the ID pin and let the firmware to control the role.
For an OTG controller, the usual method followed by the controller to assume
the role of a host or a device is governed by the state of the ID pin.

There is an alternate method where the ID pin signal is bypassed and the firmware
controls the role of the controller. In this case the USBnMODE [Bit7=IDDIG_MUX]
needs to be programmed with a value of 1 so that Register programming will be
used.

In order to work with board that have the ID pin signal and boards that have
not, this patch sets by default this register when the controller is configured
with a fixed role.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>