am37x: Adding ABB Support for 1GHz OPP
This patch adds ABB support for the AM37x 1GHz OPP.  When 1GHz
operation occurs, the ABB LDO will be enabled, effectively
boosting MPU voltage.  At any time that we are not operating
at 1GHz, the ABB LDO will be bypassed.  This change is
required for operation at 1GHz on the AM37x device.

This patch only affects AM37x.  In the course of booting,
it boosts the MPU rate to 1GHz and turns on the ABB LDO.  It
is expected that the system will be operating at 800MHz
prior to this time, as set by Uboot.  Also note that after
MPU rate is changed, the loops_per_jiffy global is updated,
which is required in order to get the proper readout from
BogoMIPS.

Upstream-Status: Not Appropriate
    * This patch is a quick fix based on an old kernel.
    * This will be reworked for the Linux mainline kernel.

Signed-off-by: Greg Guyotte <gguyotte@ti.com>