mfd: Add TWL4030 PWM driver
This PWM driver uses the PWM of the TWL4030. The driver for the TWL6030
PWM has been used as mode, but the PWM registers are different.

The driver can be used and has been tested in conjunction with
pwm-backlight to control a backlight LED of a LCD touch screen.

Signed-off-by: Bernhard Walle <walle@corscience.de>

Backported to 2.6.37 by:

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>